refactor(vm): `inspect` in separate module
