+++ crates/zksync/core/src/vm/mod.rs
@@ -1,11 +1,13 @@
 mod db;
 mod env;
 mod farcall;
+mod inspect;
 mod runner;
 mod storage_view;
 mod tracer;
 mod tx;
 
+pub use inspect::{inspect, ZKVMExecutionResult, ZKVMResult};
 pub use runner::{balance, call, code_hash, create, encode_create_params, nonce, transact};
 pub use tracer::CheatcodeTracerContext;
 pub use tx::split_tx_by_factory_deps;
